+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support to retrieve conformance classes
 - Added `Connection.makeLinksAbsolute`
+- Added `AuthProvider.getDisplayName`
 - Hook to migrate capabilities
 - Support for `usage`, `log_level`, `links` in batch jobs and services
 - Support for `level` parameter for log requests in batch jobs and services